pyproject.toml
[tool.poetry]
name = "hexproof"
version = "0.1.0"
description = "A comprehensive Magic the Gathering data and resources API."
authors = ["Investigamer <freethoughtleft@gmail.com>"]
license = "MPL-2.0"
readme = "README.md"
keywords = ["hexproof", "Magic the Gathering", "Magic", "mtg", "scryfall", "mtgjson", "python"]
classifiers = [
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.13.0"
requests = "^2.31.0"
gunicorn = "^21.2.0"
bs4 = "^0.0.1"
ratelimit = "^2.2.1"
pyyaml = "^6.0.1"
backoff = "^2.2.1"
yarl = "^1.9.3"
django-ninja = "^1.0.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
commitizen = "^3.12.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.commitizen]
version = "0.1.0"
tag_format = "v$version"
update_changelog_on_bump = true
version_files = [
    "pyproject.toml:version"
]
